// src/app/enums.ts
export enum EmojisEnum {
    GrinningFace = 'grinning-face',
    FearfulFace = 'fearful-face',
    NauseatedFace = 'nauseated-face',
    SmilingFaceWithSunglasses = 'smiling-face-with-sunglasses',
    PileOfPoop = 'pile-of-poop',
    Stopwatch = 'stopwatch',
    TriangularFlagOnPost = 'triangular-flag-on-post',
}

export enum CellCodeEnum {
    Mine = 'M',
    Flag = 'F'
}

export enum GameStatusEnum {
    Won = 'won',
    Lost = 'lost',
    Running = 'running',
    NotStarted = 'notStarted',
}

export enum GameLevelEnum {
    Easy = "easy",
    Medium = "medium",
    Hard = "hard",
}